+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 the [PEP 440 version scheme](https://peps.python.org/pep-0440/#version-scheme).
 
+## [0.5.0] - 2024-07-22
+### Added
+- `--platform-dependencies` CLI option. #8
+- `--verify/--no-verify` CLI option. #8
+
+### Removed
+- Plugin's build.yml configuration scripts. #8
+
+### Changed
+- Default value for `--platform-dependencies` CLI option to "autodetect". #9
+
 ## [0.4.0] - 2024-07-18
 ### Fixed
 - Project unusability due to outdated dependency. #6